Rewritten TermColor plugin
Rewrote the TermColor plugin to be more general instead of _only_
working with gack. You can now pass in a string argument and receive
a new formatted string as output which lets you use pretty colours
for writing to stdout manually etc.

We should probably find the exact size to use for the new string
rather than assuming a factor 4x increase but this works for now.